lots of error handling for the views

+++ campaigns/views.py
@@ -8,7 +8,10 @@ def index(request):
     return render(request, 'index.html')
 
 def campaign(request, cid):
-    campaign = Campaign.objects.get(pk=cid)
+    try:
+        campaign = Campaign.objects.get(pk=cid)
+    except:
+        return redirect('/')
     if request.user != campaign.owner:
         return redirect('/')
 
@@ -36,32 +39,56 @@ def campaign(request, cid):
     return render(request, 'campaign.html', context)
 
 def notes(request, cid):
-    campaign = Campaign.objects.get(pk=cid)
+    try:
+        campaign = Campaign.objects.get(pk=cid)
+    except:
+        return redirect('/')
+    if request.user != campaign.owner:
+        return redirect('/')
+
     threads = campaign.thread_set.all()
     sessions = campaign.session_set.all()
     events = campaign.event_set.all()
     notes = campaign.note_set.all()
+
     context = {
         'campaign': campaign,
         'threads': threads,
         'sessions': sessions,
         'notes': notes,
     }
+
     return render(request, 'notes.html', context)
 
 @ajax
 def save_event(request, eid):
-    event = Event.objects.get(pk=eid)
+    try:
+        event = Event.objects.get(pk=eid)
+    except:
+        return None
+    if request.user != event.campaign.owner:
+        return None
+
     content = request.POST.get('content')
     event.content = content
     event.save()
+
     return content
 
 @ajax
 def new_event(request, cid, sid):
-    c = Campaign.objects.get(pk=cid)
+    try:
+        c = Campaign.objects.get(pk=cid)
+    except:
+        return None
+    if request.user != c.owner:
+        return None
+
     if sid != '0':
-        s = Session.objects.get(pk=sid)
+        try:
+            s = Session.objects.get(pk=sid)
+        except:
+            return None
     else:
         lastevent = c.event_set.order_by('-time').last()
         s = lastevent.session
@@ -79,11 +106,23 @@ def new_event(request, cid, sid):
 
 @ajax
 def delete_event(request, eid):
-    event = Event.objects.get(pk=eid)
+    try:
+        event = Event.objects.get(pk=eid)
+    except:
+        return None
+    if request.user != event.campaign.owner:
+        return None
+
     event.delete()
 
 @ajax
 def note(request, nid):
-    note = Note.objects.get(pk=nid)
+    try:
+        note = Note.objects.get(pk=nid)
+    except:
+        return None
+    if request.user != note.campaign.owner:
+        return None
+
     return render(request, 'note.html', {'note': note})
